(refactor) Move topic/Search actions out of general ACTIONS
