+++ client/app/components/Feed/Search/Search.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { loadPlurbs, getPlurbs, getTopics, selectTopic } from '../../../ACTIONS.jsx';
+import { loadPlurbs, getPlurbs } from '../../../ACTIONS.jsx';
+import { getTopics, selectTopic } from './SEARCH_ACTIONS.jsx';
+
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
